+++ crates/project/src/project_graph.rs
@@ -279,7 +279,13 @@ impl ProjectGraph {
             None => return Err(ProjectError::UnconfiguredID(String::from(id))),
         };
 
-        let project = Project::new(id, source, &self.workspace_root, &self.global_config)?;
+        let project = Project::new(
+            id,
+            source,
+            &self.workspace_root,
+            &self.global_config,
+            &self.implicit_inputs,
+        )?;
         let depends_on = project.get_dependencies();
 
         // Insert the project into the graph